thats a lot of changed code :o

+++ src/main/java/com/siepert/bmnw/interfaces/IItemHazard.java
@@ -1,8 +1,8 @@
 package com.siepert.bmnw.interfaces;
 
 public interface IItemHazard {
-    default long radioactivity() {
-        return 0L;
+    default float radioactivity() {
+        return 0.0f;
     }
     default boolean burning() {
         return false;
