+++ src/main/java/com/siepert/bmnw/interfaces/IRadioactiveBlock.java
@@ -5,8 +5,8 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 public interface IRadioactiveBlock {
-    default long radioactivity(Level level, BlockPos pos, BlockState state) {
+    default float radioactivity(Level level, BlockPos pos, BlockState state) {
         return radioactivity();
     }
-    long radioactivity();
+    float radioactivity();
 }
